Separate modal into BrandModal
